refactor(editor): Move the instance-ai tools connection modal onto the registry (no-changelog)
Its two sibling modals were already registered from the descriptor while all
three key constants stayed in the shell — registration had moved, ownership
had not. All three now live in the feature.
