+++ packages/frontend/editor-ui/src/app/components/Modals.vue
@@ -36,7 +36,6 @@ import {
 	CREDENTIAL_RESOLVER_EDIT_MODAL_KEY,
 	AI_BUILDER_DIFF_MODAL_KEY,
 	INSTANCE_AI_CREDENTIAL_SETUP_MODAL_KEY,
-	INSTANCE_AI_TOOLS_CONNECTION_MODAL_KEY,
 	AI_GATEWAY_TOP_UP_MODAL_KEY,
 	TRIAL_INTRO_MODAL_KEY,
 } from '@/app/constants';
@@ -134,7 +133,6 @@ import AiGatewayTopUpModal from '@/features/ai/gateway/components/AiGatewayTopUp
 import InstanceAiCredentialSetupModal, {
 	type InstanceAiCredentialSetupModalData,
 } from '@/features/ai/instanceAi/components/InstanceAiCredentialSetupModal.vue';
-import InstanceAiToolsConnectionModalWrapper from '@/features/ai/instanceAi/components/modals/InstanceAiToolsConnectionModalWrapper.vue';
 import { defineAsyncComponent } from 'vue';
 
 const TrialIntroModal = defineAsyncComponent(
@@ -516,12 +514,6 @@ const TrialIntroModal = defineAsyncComponent(
 					:modal-name="modalName"
 					:data="data as InstanceAiCredentialSetupModalData"
 				/>
-			</template>
-		</ModalRoot>
-
-		<ModalRoot :name="INSTANCE_AI_TOOLS_CONNECTION_MODAL_KEY">
-			<template #default="{ modalName }">
-				<InstanceAiToolsConnectionModalWrapper :modal-name="modalName" />
 			</template>
 		</ModalRoot>
