+++ packages/frontend/editor-ui/src/app/stores/defaults/modals.ts
@@ -14,7 +14,6 @@ import {
 	IMPORT_CURL_MODAL_KEY,
 	IMPORT_WORKFLOW_URL_MODAL_KEY,
 	INSTANCE_AI_CREDENTIAL_SETUP_MODAL_KEY,
-	INSTANCE_AI_TOOLS_CONNECTION_MODAL_KEY,
 	LOG_STREAM_MODAL_KEY,
 	MIGRATE_WORKFLOW_MODAL_KEY,
 	NEW_ASSISTANT_SESSION_MODAL,
@@ -165,7 +164,6 @@ export const SHELL_MODAL_INITIAL_STATE: Readonly<Record<string, ModalState>> = O
 		},
 	},
 	[INSTANCE_AI_CREDENTIAL_SETUP_MODAL_KEY]: { open: false },
-	[INSTANCE_AI_TOOLS_CONNECTION_MODAL_KEY]: { open: false },
 	[INVITE_USER_MODAL_KEY]: { open: false },
 	[LOG_STREAM_MODAL_KEY]: {
 		open: false,
