+++ packages/frontend/editor-ui/src/features/ai/instanceAi/modals.ts
@@ -1,0 +1,24 @@
+import type { ModalDefinition } from '@n8n/frontend-module-sdk';
+
+import {
+	INSTANCE_AI_BROWSER_USE_SETUP_MODAL_KEY,
+	INSTANCE_AI_COMPUTER_USE_SETUP_MODAL_KEY,
+	INSTANCE_AI_TOOLS_CONNECTION_MODAL_KEY,
+} from './constants';
+
+export const INSTANCE_AI_MODALS: ModalDefinition[] = [
+	{
+		key: INSTANCE_AI_COMPUTER_USE_SETUP_MODAL_KEY,
+		component: async () => await import('./components/modals/ComputerUseSetupModal.vue'),
+	},
+	{
+		key: INSTANCE_AI_BROWSER_USE_SETUP_MODAL_KEY,
+		component: async () => await import('./components/modals/BrowserUseSetupModal.vue'),
+	},
+	{
+		key: INSTANCE_AI_TOOLS_CONNECTION_MODAL_KEY,
+		component: async () =>
+			await import('./components/modals/InstanceAiToolsConnectionModalWrapper.vue'),
+		initialState: { open: false },
+	},
+];
